% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiv-fits.r
\name{indiv_fits}
\alias{indiv_fits}
\title{Fit a model to individualized cohorts}
\usage{
indiv_fits(
  formula,
  data,
  cohorts,
  ret.model = FALSE,
  ret.augment = FALSE,
  ret.tidy = TRUE,
  ret.glance = TRUE
)
}
\arguments{
\item{formula}{A model formula, following the constraints in
\code{\link[recipes:recipe]{recipes::recipe()}}.}

\item{data}{The corpus data frame from which the \code{cohorts} were subsetted.}

\item{cohorts}{Output of \code{\link[=indiv_cohorts]{indiv_cohorts()}} (will eventually be classed).}

\item{ret.model, ret.augment, ret.tidy, ret.glance}{Logical flags for whether to
include columns for model objects or any of three tidiers (which will be
\code{NA} if methods are unavailable).}
}
\description{
An individualized model is any model fitted to an \link[=indiv_cohorts]{individualized cohort}.
}
\details{
This fitting procedure takes a model specification, a model formula,
and a slate of individualized cohorts. It fits the specified model to each
cohort, generates predictions for the corresponding new datum, and augments
the individualized cohort object with columns corresponding to the desired
objects (the model itself or any of the \link[broom:`broom-package`]{\strong{broom}}
tidiers).
}
