% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selections-list.r
\name{selections-list}
\alias{selections-list}
\alias{all_list}
\alias{all_list_predictors}
\title{Additional methods for selecting variables in step functions}
\usage{
all_list()

all_list_predictors()
}
\description{
\code{all_list()} selects columns of type 'list'. \code{all_list_predictors()} selects
'list' type columns assigned the role 'predictor'.

\strong{These need to be allowed into step functions somehow.}
}
\details{
These functions augment those documented at \code{\link[recipes:has_role]{recipes::has_role()}}.
}
